+++ cfgov/v1/__init__.py
@@ -2,12 +2,14 @@ from django.contrib.staticfiles.storage import staticfiles_storage
 from django.core.urlresolvers import reverse
 from django.template.defaultfilters import slugify
 from wagtail.wagtailcore.templatetags import wagtailcore_tags
+from django.contrib import messages
 
 import HTMLParser
 from jinja2 import Environment, contextfunction, Markup
 from sheerlike import environment as sheerlike_environment
 from compressor.contrib.jinja2ext import CompressorExtension
-
+from flags.template_functions import flag_enabled, flag_disabled
+from util.util import get_unique_id
 
 def environment(**options):
     options.setdefault('extensions', []).append(CompressorExtension)
@@ -24,18 +26,25 @@ def environment(**options):
             }
         },
         'reverse': reverse,
-        'render_stream_child': render_stream_child
+        'render_stream_child': render_stream_child,
+        'flag_enabled': flag_enabled,
+        'flag_disabled': flag_disabled,
+        'get_unique_id': get_unique_id,
+        'get_messages': messages.get_messages
     })
     env.filters.update({
         'slugify': slugify,
     })
     return env
 
-
 @contextfunction
 def render_stream_child(context, stream_child):
     # Use the django_jinja to get the template content based on its name
-    template = context.environment.get_template(stream_child.block.meta.template)
+    try:
+        template = context.environment.get_template(stream_child.block.meta.template)
+    except:
+        return stream_child
+
     # Create a new context based on the current one as we can't edit it directly
     new_context = context.get_all()
     # Add the value on the context (value is the keyword chosen by wagtail for the blocks context)