+++ cfgov/v1/forms.py
@@ -1,16 +1,58 @@
 from django import forms
+from django.forms.utils import ErrorList
 from django.forms.widgets import DateInput, SelectMultiple
-from v1.models.events import EventPage
 
+from sheerlike.query import QueryFinder
+from sheerlike.templates import date_formatter
+from v1.models.events import EventPage
+from util import ERROR_MESSAGES
+
+class FilterErrorList(ErrorList):
+    def __str__(self):
+        return '\n'.join(str(e) for e in self)
+
+class FilterDateField(forms.DateField):
+    def to_python(self, value):
+        if value:
+            try:
+                value = date_formatter(value)
+            except Exception as e:
+                pass
+        return super(FilterDateField, self).to_python(value)
+
+class FilterCheckboxList(forms.CharField):
+    def validate(self, value):
+        if value in self.empty_values and self.required:
+            msg = self.error_messages['required']
+            if self.label and '%s' in msg:
+                msg = msg % self.label
+            raise forms.ValidationError(msg, code='required')
 
 class CalenderPDFFilterForm(forms.Form):
-    filter_calendar = forms.CharField()
-    filter_range_date_gte = forms.DateField(input_formats=['%Y-%m-%d'])
-    filter_range_date_lte = forms.DateField(input_formats=['%Y-%m-%d'])
+    filter_calendar = FilterCheckboxList(label='Calendar',
+        error_messages=ERROR_MESSAGES['CHECKBOX_ERRORS'])
+    filter_range_date_gte = FilterDateField(required=False,
+        error_messages=ERROR_MESSAGES['DATE_ERRORS'])
+    filter_range_date_lte = FilterDateField(required=False,
+        error_messages=ERROR_MESSAGES['DATE_ERRORS'])
+
+    def __init__(self, *args, **kwargs):
+        kwargs['error_class'] = FilterErrorList
+        super(CalenderPDFFilterForm, self).__init__(*args, **kwargs)
 
     def clean_filter_calendar(self):
         return self.cleaned_data['filter_calendar'].replace(' ', '+')
 
+    def clean(self):
+        cleaned_data = super(CalenderPDFFilterForm, self).clean()
+        from_date_empty = 'filter_range_date_gte' in cleaned_data and \
+                          cleaned_data['filter_range_date_gte']  == None
+        to_date_empty = 'filter_range_date_lte' in cleaned_data and \
+                        cleaned_data['filter_range_date_lte'] == None
+
+        if from_date_empty and to_date_empty :
+            raise forms.ValidationError(ERROR_MESSAGES['DATE_ERRORS']['one_required'])
+        return cleaned_data
 
 class EventsFilterForm(forms.Form):
     tags_select_attrs = {
@@ -34,7 +76,6 @@ class EventsFilterForm(forms.Form):
                               widget=DateInput(attrs=to_select_attrs))
     topics = forms.MultipleChoiceField(required=False, choices=[],
                                        widget=SelectMultiple(attrs=tags_select_attrs))
-
     @property
     def field_queries(self):
         return zip(['start_dt__gte', 'end_dt__lte', 'tags__name__in'],
@@ -42,6 +83,7 @@ class EventsFilterForm(forms.Form):
 
     def __init__(self, *args, **kwargs):
         super(EventsFilterForm, self).__init__(*args, **kwargs)
+
         if 'topics' in self.fields:
             options = list(set([tag for tags in [event.tags.names() for event
                                                  in EventPage.objects.live()]