+++ cfgov/v1/models/base.py
@@ -1,7 +1,4 @@
 import os
-import sys
-
-from elasticsearch.exceptions import NotFoundError
 
 from django.db import models
 from django.http import Http404
@@ -10,19 +7,23 @@ from django.utils.translation import ugettext_lazy as _
 from wagtail.wagtailadmin.edit_handlers import StreamFieldPanel
 from wagtail.wagtailcore import blocks
 from wagtail.wagtailcore.fields import StreamField
-from wagtail.wagtailcore.models import Page, PagePermissionTester, UserPagePermissionsProxy, Orderable
+from wagtail.wagtailcore.models import Page, PagePermissionTester, \
+    UserPagePermissionsProxy, Orderable
 from wagtail.wagtailcore.url_routing import RouteResult
-from wagtail.wagtailadmin.edit_handlers import FieldPanel, InlinePanel, MultiFieldPanel, TabbedInterface, ObjectList
+from wagtail.wagtailadmin.edit_handlers import FieldPanel, InlinePanel, \
+    MultiFieldPanel, TabbedInterface, ObjectList
 
 from taggit.models import TaggedItemBase
 from modelcluster.fields import ParentalKey
 from modelcluster.tags import ClusterTaggableManager
 
-from sheerlike.query import get_document, more_like_this
+from sets import Set
+
 from . import ref
 from . import atoms
 from . import molecules
 from . import organisms
+from ..util import util
 
 
 class CFGOVAuthoredPages(TaggedItemBase):
@@ -67,7 +68,7 @@ class CFGOVPage(Page):
         ('contact', organisms.MainContactInfo()),
     ], blank=True)
 
-    ### Panels ###
+    # Panels
     sidefoot_panels = [
         StreamFieldPanel('sidefoot'),
     ]
@@ -130,7 +131,9 @@ class CFGOVPage(Page):
                 return _("live")
 
     def sharable_pages(self):
-        """Return a queryset of the pages that this user has permission to share"""
+        """
+        Return a queryset of the pages that this user has permission to share.
+        """
         # Deal with the trivial cases first...
         if not self.user.is_active:
             return Page.objects.none()
@@ -142,7 +145,8 @@ class CFGOVPage(Page):
         for perm in self.permissions.filter(permission_type='share'):
             # User has share permission on any subpage of perm.page
             # (including perm.page itself).
-            sharable_pages |= Page.objects.descendant_of(perm.page, inclusive=True)
+            sharable_pages |= Page.objects.descendant_of(perm.page,
+                                                         inclusive=True)
 
         return sharable_pages
 
@@ -173,13 +177,42 @@ class CFGOVPage(Page):
 
     def permissions_for_user(self, user):
         """
-        Return a CFGOVPagePermissionTester object defining what actions the user can perform on this page
+        Return a CFGOVPagePermissionTester object defining what actions the
+        user can perform on this page
         """
         user_perms = CFGOVUserPagePermissionsProxy(user)
         return user_perms.for_page(self)
 
     class Meta:
         app_label = 'v1'
+
+    def elements(self):
+        return []
+
+    def _media(self):
+        from v1 import models
+
+        js = ()
+
+        for child in self.elements():
+            if isinstance(child, dict):
+                type = child['type']
+            else:
+                type = child[0]
+
+            class_ = getattr(models, util.to_camel_case(type))
+
+            instance = class_()
+
+            try:
+                if hasattr(instance.Media, 'js'):
+                    js += instance.Media.js
+            except:
+                pass
+
+        return Set(js)
+
+    media = property(_media)
 
 
 class CFGOVPageCategory(Orderable):